requirements.txt
django-autofixture==0.12.0
flake8==2.5.4
django==1.9.4
django-extensions==1.6.1
django-filter==0.12.0
djangorestframework==3.3.2
psycopg2==2.6.1
whitenoise==3.3.1
gunicorn==19.4.5
dj-database-url==0.4.0
